requirements.txt
streamlit
streamlit-authenticator
pymilvus
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pyyaml
alfred-vault[all]==0.3.1
